replace waitForMessage with observable
